# src/test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: support-management
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "12.4"
servers:
  - url: http://localhost:51578
    description: Generated server url
...
paths:
...
  /{municipalityId}/{namespace}/metadata/statuses:
    get:
      tags:
        - Metadata for statuses
      summary: Get statuses
      description: Get all statuses for the namespace and municipality
      operationId: getStatuses
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Status"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
        - Metadata for statuses
      summary: Create status
      description: Create new status for the namespace and municipality
      operationId: createStatus
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Status"
        required: true
      responses:
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/metadata/statuses/{status}:
    get:
      tags:
        - Metadata for statuses
      summary: Get status
      description: "Get status matching sent in namespace, municipality and status"
      operationId: getStatus
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: status
          in: path
          description: Name of status
          required: true
          schema:
            type: string
          example: NEW
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
        - Metadata for statuses
      summary: Delete status
      description: "Delete status matching namespace, municipality and status"
      operationId: deleteStatus
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: status
          in: path
          description: Name of status
          required: true
          schema:
            type: string
          example: PRIVATE
      responses:
        "204":
          description: Successful operation
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: { }
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    Status:
      type: object
      description: Status model
      properties:
        name:
          type: string
          description: Name for the status
          examples:
            - statusName
          minLength: 1
        created:
          type: string
          format: date-time
          description: Timestamp when the status was created
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
        modified:
          type: string
          format: date-time
          description: Timestamp when the status was last modified
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
      required:
        - name
...
  securitySchemes: { }